New sugar detection system for agent.
New approach: Model sugar as a radial zone. Have agent detect first collision with sugar and from that point on check the sugar concentration at co-ord. Sugar concentration still needs to be mapped on to the radial zone. Currently using a GameObject created in the Unity3d visual editor. Need to move to programmatically creating the SugarGradient GameObject and map concentration values.