More accurate chemical sampling.
E. coli can now sample any part of the environment and get a chemical sample based on the diffusion of all chemicals present in the agar. Currently the highest concentrated instance of a chemical becomes the base chemical for samples. Will work on changing that so that the sample obtained from any point is the net concentration of all chemicals in the Agar, calculated by diffusion from Chemicals origin point.