Swarm intelligence emerging.
E. coli releasing of Attractant that has a concentration based on the E. coli's last Attractant measurement is resulting in a swarm of E. coli that seem to act as one body with a center of mass. The pull to this center of mass might be currently too strong as the Attractants emmited by the E. coli are unlimited. Need to have a trade off between staying near the center of mass and exploring further. Could be metabolic based.